refactor: fix namespaces for relocated and renamed override classes
