+++ tests/Unit/Managers/ServiceOverrideManagerTest.php
@@ -7,8 +7,8 @@ use PHPUnit\Framework\Attributes\Test;
 use Sprout\Core\Exceptions\MisconfigurationException;
 use Sprout\Core\Exceptions\TenancyMissingException;
 use Sprout\Core\Managers\ServiceOverrideManager;
-use Sprout\Core\Overrides\CookieOverride;
-use Sprout\Core\Overrides\SessionOverride;
+use Sprout\Overrides\Cookie\CookieOverride;
+use Sprout\Overrides\Session\SessionOverride;
 use Sprout\Core\TenancyOptions;
 use Sprout\Core\Tests\Unit\UnitTestCase;
 use Workbench\App\Models\TenantModel;
