+++ tests/Unit/Overrides/Broadcast/BudBroadcastOverrideTest.php
@@ -18,8 +18,8 @@ use Sprout\Bud\Contracts\ConfigStore;
 use Sprout\Bud\Exceptions\CyclicOverrideException;
 use Sprout\Bud\Managers\ConfigStoreManager;
 use Sprout\Bud\Overrides\Broadcast\BudBroadcastManager;
-use Sprout\Bud\Overrides\BroadcastConnectionOverride;
-use Sprout\Bud\Overrides\BroadcastManagerOverride;
+use Sprout\Overrides\Broadcast\BudBroadcastConnectionOverride;
+use Sprout\Overrides\Broadcast\BudBroadcastManagerOverride;
 use Sprout\Bud\Tests\Unit\UnitTestCase;
 use Sprout\Core\Contracts\BootableServiceOverride;
 use Sprout\Core\Contracts\Tenancy;
