+++ tests/Unit/Overrides/Cache/BudCacheStoreOverrideTest.php
@@ -15,7 +15,7 @@ use Sprout\Bud\Bud;
 use Sprout\Bud\Contracts\ConfigStore;
 use Sprout\Bud\Exceptions\CyclicOverrideException;
 use Sprout\Bud\Managers\ConfigStoreManager;
-use Sprout\Bud\Overrides\CacheStoreOverride;
+use Sprout\Overrides\Cache\BudCacheStoreOverride;
 use Sprout\Bud\Tests\Unit\UnitTestCase;
 use Sprout\Core\Contracts\BootableServiceOverride;
 use Sprout\Core\Contracts\Tenancy;
