+++ tests/Unit/Overrides/Filesystem/BudFilesystemDiskOverrideTest.php
@@ -17,14 +17,14 @@ use Sprout\Bud\Bud;
 use Sprout\Bud\Contracts\ConfigStore;
 use Sprout\Bud\Exceptions\CyclicOverrideException;
 use Sprout\Bud\Managers\ConfigStoreManager;
-use Sprout\Bud\Overrides\FilesystemDiskOverride;
+use Sprout\Overrides\Filesystem\BudFilesystemDiskOverride;
 use Sprout\Bud\Tests\Unit\UnitTestCase;
 use Sprout\Core\Contracts\BootableServiceOverride;
 use Sprout\Core\Contracts\Tenancy;
 use Sprout\Core\Contracts\Tenant;
 use Sprout\Core\Contracts\TenantHasResources;
 use Sprout\Core\Overrides\Filesystem\SproutFilesystemManager;
-use Sprout\Core\Overrides\FilesystemManagerOverride;
+use Sprout\Overrides\Filesystem\FilesystemManagerOverride;
 use Sprout\Core\Overrides\StackedOverride;
 use Sprout\Core\Sprout;
 use Sprout\Core\Support\SettingsRepository;
