+++ tests/Unit/Overrides/Filesystem/FilesystemOverrideTest.php
@@ -15,8 +15,8 @@ use Sprout\Core\Contracts\BootableServiceOverride;
 use Sprout\Core\Contracts\Tenancy;
 use Sprout\Core\Contracts\Tenant;
 use Sprout\Core\Contracts\TenantHasResources;
-use Sprout\Core\Overrides\FilesystemManagerOverride;
-use Sprout\Core\Overrides\FilesystemOverride;
+use Sprout\Overrides\Filesystem\FilesystemManagerOverride;
+use Sprout\Overrides\Filesystem\FilesystemOverride;
 use Sprout\Core\Overrides\StackedOverride;
 use Sprout\Core\Sprout;
 use Sprout\Core\Support\SettingsRepository;
