+++ tests/Unit/Overrides/StackedOverrideTest.php
@@ -7,8 +7,8 @@ use Illuminate\Contracts\Foundation\Application;
 use Mockery\MockInterface;
 use PHPUnit\Framework\Attributes\Test;
 use Sprout\Core\Exceptions\MisconfigurationException;
-use Sprout\Core\Overrides\AuthGuardOverride;
-use Sprout\Core\Overrides\AuthPasswordOverride;
+use Sprout\Overrides\Auth\AuthGuardOverride;
+use Sprout\Overrides\Auth\AuthPasswordOverride;
 use Sprout\Core\Overrides\StackedOverride;
 use Sprout\Core\Sprout;
 use Sprout\Core\Support\SettingsRepository;
